move `extractRemoteSafeInfo` to `utils/extractRemoteSafeInfo` and `extractSafeOwners` to `utils/buildSafeOwners`
